+++ CHANGELOG.md
@@ -21,6 +21,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   source tree (#154).
 - Support distribution and format specific tokens of prescript rules to allow
   definition of different values for different distributions and formats (#156).
+- Support installation of DNF modules as prescript dependencies for RPM packages
+  builds (#155).
 - cli:
   - Possibility to list artifacts in registries in remote instance with the REST
     API (#141)